% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Annotate the tree}
\usage{
annotate_tree(tree, tree_data, get_sp, data_sp_column)
}
\arguments{
\item{tree}{Phylogenetic tree imported with the treeio::read.tree().}

\item{tree_data}{Dataframe containing information for the species in the tree.}

\item{get_sp}{function to extract the species code from the tip.label.}

\item{data_sp_column}{Column containing the species code.}
}
\value{
A phylo4 tree object with a data frame with the annotations.
}
\description{
Function to annotate the tree with the data given in the tree information
dataframe 'data'. This dataframe must contain one column with the species
code matching the species code in the headers which is extracted by the
'get_sp' function
}
\examples{
library(treeio)

get_sp <- function(header) {
  return(header)
}

set.seed(2024-11-04)
tree <- rtree(20)

tree_data <- data.frame(tip = tree$tip.label,
                        group = c(rep('A', 10), rep('B', 10)))
tree_data

atree <- annotate_tree(tree = tree,
                       tree_data = tree_data,
                       get_sp = get_sp,
                       data_sp_column = 1)

atree
}
\author{
Moisès Bernabeu
}
