% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_monophyletics}
\alias{get_monophyletics}
\title{Get monophyletic groups}
\usage{
get_monophyletics(
  tree,
  group_column,
  tree_data = NULL,
  get_sp = NULL,
  data_sp_column = NULL
)
}
\arguments{
\item{tree}{Phylogenetic tree.}

\item{group_column}{Column to extract the monophyletic groups.}

\item{tree_data}{Dataframe containing information for the species in the tree.}

\item{get_sp}{Function to extract the species code from the tip.label.}

\item{data_sp_column}{Column containing the species code.}
}
\value{
Vector with the sister nodes.
}
\description{
Get the monophyletic groups assigned to a specific column.
}
\examples{
library(treeio)
library(ggtree)

get_sp <- function(header) {
  return(header)
}

set.seed(2024-11-04)
tree <- rtree(20)

tree_data <- data.frame(tip = tree$tip.label,
                        group = c(rep('A', 10), rep('B', 10)))
atree <- annotate_tree(tree = tree,
                       tree_data = tree_data,
                       get_sp = get_sp,
                       data_sp_column = 1)

mphy <- get_monophyletics(atree, 'group')

mphy

ggtree(atree) +
  geom_nodelab(aes(subset = node \%in\% mphy, label = node), geom = 'label') +
  geom_tiplab(aes(label = group))
}
\author{
Moisès Bernabeu
}
