% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_sisters}
\alias{get_sisters}
\title{Get sisters}
\usage{
get_sisters(tree, node, naming_column = NULL)
}
\arguments{
\item{tree}{Phylogenetic tree.}

\item{node}{The number of name (in case of a tip) to start extracting sister nodes.}

\item{naming_column}{Column to name the sister groups based on the proportion of the column values among the children tips.}
}
\value{
integer: vector with the sister nodes.
}
\description{
Get the number of the ancestor of the sister nodes to a specified node.
}
\details{
Args:
}
\examples{
library(treeio)
library(ggtree)

get_sp <- function(header) {
  return(header)
}

set.seed(2024-11-04)
tree <- rtree(20)

tree_data <- data.frame(tip = tree$tip.label,
                        group = c(rep('A', 10), rep('B', 10)))
atree <- annotate_tree(tree = tree,
                       tree_data = tree_data,
                       get_sp = get_sp,
                       data_sp_column = 1)

ggtree(atree) +
  geom_nodelab(aes(label = node), geom = 'label') +
  geom_tiplab()

# Getting sisters to node 36
get_sisters(tree, 36)

# Getting sisters to t7
get_sisters(atree, 't7')

# Getting sisters to node 36 annotated with the percentage of tips with a
# certain value of the given naming_column
get_sisters(atree, 36, naming_column = 'group')
}
\author{
Moisès Bernabeu
}
