% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcollapse.R
\name{ggcollapse}
\alias{ggcollapse}
\title{Plot a collapsed tree}
\usage{
ggcollapse(
  tree,
  nodes,
  collapse_mode = "mixed",
  node_colours = NULL,
  get_sp = NULL,
  tree_data = NULL,
  data_sp_column = NULL
)
}
\arguments{
\item{tree}{Phylogenetic tree.}

\item{nodes}{Named vector conaining the nodes to collapse.}

\item{collapse_mode}{String to select the aesthetics of the collapsing triangle, 'mixed', 'min' or 'max'.}

\item{node_colours}{Vector named with the same names as the nodes with the colour codes.}

\item{get_sp}{Function to extract the species from the tip.label.}

\item{tree_data}{Dataframe with the data linked to each species or sequence.}

\item{data_sp_column}{Column of the tree_data containing the species code.}
}
\value{
Collapsed tree ggplot plot object.
}
\description{
This function takes the nodes to collapse, prunes all the leaves except the
longest and the shortest, collapse them and assign some colours and names
according to the arguments given.
}
\examples{
library(treeio)
library(ggtree)

get_sp <- function(header) {
  return(header)
}

set.seed(2024-11-04)
tree <- rtree(20)

tree_data <- data.frame(tip = tree$tip.label,
                        group = c(rep('A', 10), rep('B', 10)))
atree <- annotate_tree(tree = tree,
                       tree_data = tree_data,
                       get_sp = get_sp,
                       data_sp_column = 1)

# Collapsing monophyletics of the group
mphy <- get_monophyletics(atree, 'group')
cols <- c('A' = 'steelblue', 'B' = 'darkorange')

ggcollapse(tree, mphy, collapse_mode = 'mixed', node_colours = cols)

# Collapsing sisters
sist <- get_sisters(atree, 36, 'group')

ggcollapse(tree, sist, collapse_mode = 'mixed') +
  hexpand(0.15)
}
\author{
Moisès Bernabeu
}
